requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
greenlet==3.0.3
idna==3.7
numpy==1.26.4
packaging==24.0
pandas==2.0.3
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
SQLAlchemy==1.4.52
sqlalchemy-redshift==0.8.14
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1